refactor: move createClient in its own file and rename defaultProps to plainClientDefaultProps
