+++ lib/plain/plain-client-types.ts
@@ -120,13 +120,13 @@ import type { WebhookPlainClientAPI } from './entities/webhook'
 import type { WorkflowPlainClientAPI } from './entities/workflow'
 import type { WorkflowDefinitionPlainClientAPI } from './entities/workflow-definition'
 import type { WorkflowsChangelogPlainClientAPI } from './entities/workflows-changelog'
-import type { DefaultParams, OptionalDefaults } from './wrappers/wrap'
+import type { PlainClientDefaultParams, OptionalDefaults } from './wrappers/wrap'
 import type { OAuthApplicationPlainClientAPI } from './entities/oauth-application'
 import type { FunctionLogPlainClientAPI } from './entities/function-log'
 
 export type PlainClientAPI = {
   raw: {
-    getDefaultParams(): DefaultParams | undefined
+    getDefaultParams(): PlainClientDefaultParams | undefined
     get<T = unknown>(url: string, config?: RawAxiosRequestConfig): Promise<T>
     post<T = unknown>(url: string, payload?: any, config?: RawAxiosRequestConfig): Promise<T>
     patch<T = unknown>(url: string, payload?: any, config?: RawAxiosRequestConfig): Promise<T>
